Refactor: Mv `process_spawned_*` into `process_change_of_runtime.rs`
